Fix how unsharing is processed
That version don't use the option --unshare-all
of bubblewrap anymore because it didn't fit the
expected requirement. For example, if the configuration
of the rednode was telling:

 share_all: Enabled
 share_cgroup: Disabled

then bublewrap was called with option

 --unshare-all

What was not the expected behaviour.

Now with that fix, bubblewrap is called with options

 --unshare-cgroup

as expected.